Refactoring out the Reporter from the step call
Created a new method, runUnreportedStep, that simply grabs the step out of a match,
throwing an exception if it does not match, and executes it.

All the exceptions are required to trickle up so that they're caught by the step that
does the actual execution so that things look right in the output.

A snippit will be generated if you call a non existing step, so that's handy :)