[Build] Improve method by which examples are included from deployment
When using local staging the deployment is deferred to the last module.
Because half the project consists of examples that should not be
deployed and thus skip the execution of the nexus staging plugin
this causes the whole whole project to skip deployment.

By having `continuous-deployment` as last module in the build we ensured
that the nexus plugin is triggered uploads all artifacts at the end of
the build.

This however results in the inclusion of the `continuous-deployment`
artifact. By putting examples in a profile that is active by default
we do ensure it will always be build and but can exclude it from the
actual release.